src/requirements.txt
python-dotenv==1.0.1
huggingface-hub==0.29.3
langchain==0.3.20
langchain-huggingface==0.1.2
sentence-transformers==3.4.1
llama-index==0.12.25
gradio==5.22.0